+++ addon/classes/Row.js
@@ -1,11 +1,19 @@
 import Ember from 'ember';
 
+const { computed, guidFor } = Ember;
+
 /**
  * @module Table
+ * @extends Ember.ObjectProxy
  * @class Row
  */
 export default class Row extends Ember.ObjectProxy.extend({
   /**
+   * Whether the row is hidden.
+   *
+   * CSS Classes:
+   *  - `is-hidden`
+   *
    * @property hidden
    * @type {Boolean}
    * @default false
@@ -13,6 +21,11 @@ export default class Row extends Ember.ObjectProxy.extend({
   hidden: false,
 
   /**
+   * Whether the row is expanded.
+   *
+   * CSS Classes:
+   *  - `is-expanded`
+   *
    * @property expanded
    * @type {Boolean}
    * @default false
@@ -20,6 +33,11 @@ export default class Row extends Ember.ObjectProxy.extend({
   expanded: false,
 
   /**
+   * Whether the row is selected.
+   *
+   * CSS Classes:
+   *  - `is-selected`
+   *
    * @property selected
    * @type {Boolean}
    * @default false
@@ -32,7 +50,32 @@ export default class Row extends Ember.ObjectProxy.extend({
    * @property classNames
    * @type {String | Array}
    */
-  classNames: null
+  classNames: null,
+
+  /**
+   * Data content for this row. Since this class extends Ember.ObjectProxy,
+   * all properties are forwarded to the content. This means that instead of
+   * `row.content.foo` you can just do `row.foo`. Please note that methods are
+   * not forwarded. You will not be able to do `row.save()`, instead, you would have
+   * to do `row.content.save()`.
+   *
+   * @property content
+   * @type {Object}
+   */
+  content: null,
+
+  /**
+   * Rows's unique ID.
+   *
+   * Note: named `rowId` in order to not shadow the `content.id` property.
+   *
+   * @property rowId
+   * @type {String}
+   * @readOnly
+   */
+  rowId: computed(function() {
+    return guidFor(this);
+  }).readOnly()
 }) {
   /**
    * @class Row