+++ tests/dummy/app/breakpoints.js
@@ -1,0 +1,6 @@
+export default {
+  mobile:  '(max-width: 768px)',
+  tablet:  '(min-width: 769px) and (max-width: 992px)',
+  desktop: '(min-width: 993px) and (max-width: 1200px)',
+  jumbo:   '(min-width: 1201px)'
+};